+++ src/index.ts
@@ -1,14 +1,39 @@
 import './util/androidFetch'
 
+// ---------------------------------------------------------------------
+// Components
+// ---------------------------------------------------------------------
+
 export { LoginUiProvider } from './components/publicApi/LoginUiProvider'
-export * from './components/publicApi/ChangePasswordScreen'
-export * from './components/publicApi/ChangePinScreen'
-export * from './components/publicApi/UpgradeUsernameScreen'
-export * from './components/publicApi/ChangeRecoveryScreen'
-export * from './components/publicApi/LoginScreen'
-export * from './components/publicApi/OtpRepairScreen'
-export * from './components/publicApi/SecurityAlertsScreen'
-export * from './components/publicApi/types'
+export { ChangePasswordScreen } from './components/publicApi/ChangePasswordScreen'
+export { ChangePinScreen } from './components/publicApi/ChangePinScreen'
+export { UpgradeUsernameScreen } from './components/publicApi/UpgradeUsernameScreen'
+export { PasswordRecoveryScreen } from './components/publicApi/ChangeRecoveryScreen'
+export { LoginScreen } from './components/publicApi/LoginScreen'
+export { OtpRepairScreen } from './components/publicApi/OtpRepairScreen'
+export { SecurityAlertsScreen } from './components/publicApi/SecurityAlertsScreen'
+export type {
+  AppConfig,
+  CreateAccountType,
+  InitialRouteName,
+  NotificationPermissionReminderOptions
+} from './components/publicApi/publicTypes'
+
+// ---------------------------------------------------------------------
+// Post-login steps
+// ---------------------------------------------------------------------
+
+export { refreshTouchId } from './keychain'
+export { showNotificationPermissionReminder } from './util/notificationPermissionReminder'
+export { showOtpReminder } from './util/otpReminder'
+export {
+  hasSecurityAlerts,
+  watchSecurityAlerts
+} from './util/hasSecurityAlerts'
+
+// ---------------------------------------------------------------------
+// Biometric utilities
+// ---------------------------------------------------------------------
 
 export {
   disableTouchId,
@@ -16,8 +41,3 @@ export {
   isTouchEnabled,
   getSupportedBiometryType
 } from './keychain'
-
-export {
-  hasSecurityAlerts,
-  watchSecurityAlerts
-} from './util/hasSecurityAlerts'